Refactor UI schema handling: - Moved from ComputeBuilder.Plugin.Models to ComputeBuilder.Plugin.Models.Generated. - Changed GrasshopperParamType from enum to string for compatibility. - Updated related files and components to reflect the new schema structure. - Removed obsolete UISchema.cs file and updated TypeScript imports accordingly. - Enhanced layout item handling with if-else structure for clarity. - Added comprehensive README for schema generation and type management.

+++ Plugin/Utils/ParameterTypeHelper.cs
@@ -1,5 +1,5 @@
 using System;
-using ComputeBuilder.Plugin.Models;
+using ComputeBuilder.Plugin.Models.Generated;
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Special;
 
@@ -75,7 +75,7 @@ namespace ComputeBuilder.Plugin.Utils
                 availableParam.Maximum = maximum.Value;
                 if (stepSize.HasValue)
                 {
-                    availableParam.StepSize = stepSize.Value;
+                    availableParam.StepSize = (double)stepSize.Value;
                 }
             }
             else
